# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Welcome to the HavenStay Backend API documentation! 🚀\n\nHavenStay is a modern, high-performance Domain-Driven Design (DDD) backend built in Go. It empowers Hosts to manage luxury properties and Guests to seamlessly book remote getaways.\n\n### Key Features:\n- **🏘️ Property Management**: Full CRUD lifecycle for short-term rentals and luxury homes.\n- **👩‍💼 Host \u0026 Guest Profiles**: Independent onboarding and profile management domains.\n- **📅 Booking Engine**: Real-time reservation creation.\n- **❤️ Wishlists**: Guests can curate lists of favorite properties.\n\n**Built With**: Go 1.21+, Gin (REST), gRPC, Protobufs, MySQL 8, GORM.",
        "title": "HavenStay - Hotel \u0026 Room Booking API",
        "termsOfService": "https://havenstay.localhost/terms",
        "contact": {
            "name": "HavenStay Platform Support",
            "url": "https://github.com/AshutoshKY/havenstay-backend",
            "email": "api-support@havenstay.local"
        },
        "license": {
            "name": "MIT License",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/v1/hosts": {
            "post": {
                "description": "Creates a new Host profile on the platform. Hosts are independent users capable of listing and managing multiple Properties.\nNote: A newly created host is automatically marked as `is_verified: false` pending administrator approval.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Hosts"
                ],
                "summary": "Register a new Host",
                "parameters": [
                    {
                        "description": "Payload containing the Host's basic contact and profile information (Name, Email, Phone, Location).",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_user_airbnb-test_api_proto_v1.CreateHostRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Host successfully registered and assigned a unique ID.",
                        "schema": {
                            "$ref": "#/definitions/github_com_user_airbnb-test_api_proto_v1.HostResponse"
                        }
                    },
                    "400": {
                        "description": "Validation Error - Invalid email format or missing required fields.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error - Database connection failure.",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/hosts/{id}": {
            "get": {
                "description": "Fetches the complete public profile of a Host by their unique identifier natively from the core system.\nUse this endpoint to display a Host's verification status, contact info, and join date on their public landing page.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Hosts"
                ],
                "summary": "Retrieve Host details",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "The unique numeric ID of the Host to retrieve.",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The requested Host details were found and returned successfully.",
                        "schema": {
                            "$ref": "#/definitions/github_com_user_airbnb-test_api_proto_v1.HostResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request - The provided ID was not a valid integer.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found - No Host exists with the provided ID.",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/v1/properties": {
            "get": {
                "description": "Retrieve a sweeping list of all active property listings on the HavenStay platform.\n*(Note: In future versions, this will include global pagination and bounding-box map filters).*",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Properties"
                ],
                "summary": "Browse all Properties",
                "responses": {
                    "200": {
                        "description": "A comprehensive array of all available properties.",
                        "schema": {
                            "$ref": "#/definitions/github_com_user_airbnb-test_api_proto_v1.ListPropertiesResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error while querying the database.",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "An authenticated Host can use this endpoint to publish a new real-estate listing.\nNote: You MUST pass a valid `host_id` representing an already-registered Host on the platform, or the database foreign-key constraint will reject the listing.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Properties"
                ],
                "summary": "List a new Property",
                "parameters": [
                    {
                        "description": "The core details of the listing: Name, Description, Location String, and Nightly Price (Float).",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_user_airbnb-test_api_proto_v1.CreatePropertyRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Property was published and is now live on the index.",
                        "schema": {
                            "$ref": "#/definitions/github_com_user_airbnb-test_api_proto_v1.PropertyResponse"
                        }
                    },
                    "400": {
                        "description": "Validation Bad Request - Usually missing Host ID or malformed price.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error.",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/properties/host/{hostId}": {
            "get": {
                "description": "Get all properties listed specifically by one Host.\nThis endpoint powers the \"Host Dashboard\", allowing property owners to see all assets they have under management. Optionally, you can pass a `location` query to filter the host's properties by a specific city.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Properties"
                ],
                "summary": "Host's Dashboard Properties",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "The Host ID managing these properties.",
                        "name": "hostId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Optional string to filter the resulting list by a specific city or region.",
                        "name": "location",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list of properties bound to the specified host.",
                        "schema": {
                            "$ref": "#/definitions/github_com_user_airbnb-test_api_proto_v1.ListPropertiesResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error.",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/properties/{id}": {
            "get": {
                "description": "Drill down into a specific property. Returns the exact title, host bindings, textual description, nightly price, and location coordinates formatted for front-end rendering.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Properties"
                ],
                "summary": "Get Property Details",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "The exact Property ID to inspect.",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Full property overview returned.",
                        "schema": {
                            "$ref": "#/definitions/github_com_user_airbnb-test_api_proto_v1.PropertyResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Property Not Found - the listing may have been deactivated.",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "github_com_user_airbnb-test_api_proto_v1.CreateHostRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "user_name": {
                    "type": "string"
                }
            }
        },
        "github_com_user_airbnb-test_api_proto_v1.CreatePropertyRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "host_id": {
                    "type": "integer"
                },
                "location": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_user_airbnb-test_api_proto_v1.Host": {
            "type": "object",
            "properties": {
                "created_at": {
                    "$ref": "#/definitions/timestamppb.Timestamp"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_verified": {
                    "type": "boolean"
                },
                "location": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "updated_at": {
                    "$ref": "#/definitions/timestamppb.Timestamp"
                },
                "user_name": {
                    "type": "string"
                }
            }
        },
        "github_com_user_airbnb-test_api_proto_v1.HostResponse": {
            "type": "object",
            "properties": {
                "host": {
                    "$ref": "#/definitions/github_com_user_airbnb-test_api_proto_v1.Host"
                }
            }
        },
...
        "github_com_user_airbnb-test_api_proto_v1.ListPropertiesResponse": {
            "type": "object",
            "properties": {
                "properties": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_user_airbnb-test_api_proto_v1.Property"
                    }
                }
            }
        },
        "github_com_user_airbnb-test_api_proto_v1.Property": {
            "type": "object",
            "properties": {
                "created_at": {
                    "$ref": "#/definitions/timestamppb.Timestamp"
                },
                "description": {
                    "type": "string"
                },
                "host_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "location": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                }
            }
        },
        "github_com_user_airbnb-test_api_proto_v1.PropertyResponse": {
            "type": "object",
            "properties": {
                "property": {
                    "$ref": "#/definitions/github_com_user_airbnb-test_api_proto_v1.Property"
                }
            }
        },
...
        "timestamppb.Timestamp": {
            "type": "object",
            "properties": {
                "nanos": {
                    "description": "Non-negative fractions of a second at nanosecond resolution. This field is\nthe nanosecond portion of the duration, not an alternative to seconds.\nNegative second values with fractions must still have non-negative nanos\nvalues that count forward in time. Must be between 0 and 999,999,999\ninclusive.",
                    "type": "integer"
                },
                "seconds": {
                    "description": "Represents seconds of UTC time since Unix epoch 1970-01-01T00:00:00Z. Must\nbe between -315576000000 and 315576000000 inclusive (which corresponds to\n0001-01-01T00:00:00Z to 9999-12-31T23:59:59Z).",
                    "type": "integer"
                }
            }
        }
    }
}